backend/src/newsletter/serializers.py
from rest_framework import serializers

from newsletter.models import NewsletterSubscription


class NewsletterSubscriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = NewsletterSubscription
        fields = ["id", "email", "created_at"]
